rename all .h -> .hpp, include tidy-up
