+++ winter/Code/binding/register_types.cpp
@@ -3,7 +3,7 @@
 #include <godot_cpp/core/class_db.hpp>
 #include <godot_cpp/core/defs.hpp>
 #include <godot_cpp/godot.hpp>
-#include "binding.h"
+#include "binding.hpp"
 
 using namespace godot;
