+++ winter/Code/emley/CMakeLists.txt
@@ -4,7 +4,7 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 find_package(CURL REQUIRED)
 FetchContent_Declare(json URL https://github.com/nlohmann/json/releases/download/v3.11.2/json.tar.xz)
 FetchContent_MakeAvailable(json)
-add_library(${PROJECT_NAME} kachemak/kachemak.cpp version/version.cpp )
+add_library(${PROJECT_NAME} kachemak/kachemak.cpp version/version.cpp)
 target_include_directories(emley PUBLIC ./)
 target_include_directories(emley PUBLIC ../shared)
 target_include_directories(emley PUBLIC ../net)
