+++ winter/Code/headless/headless.hpp
@@ -1,4 +1,4 @@
 #pragma once
-#include <adastral_defs.h>
+#include <adastral_defs.hpp>
 #include <palace.hpp>
 #define WELCOME_TEXT "Adastral (headless), git build. (c) 2024 The Adastral Group and contributors.\n"
