+++ winter/Code/palace/CMakeLists.txt
@@ -2,7 +2,7 @@ project(palace LANGUAGES C CXX)
 find_package(CURL REQUIRED)
 # Get the name of this folder as the project name
 get_filename_component(PROJECT_NAME ${CMAKE_CURRENT_SOURCE_DIR} NAME_WE)
-add_library(${PROJECT_NAME} palace.cpp )
+add_library(${PROJECT_NAME} palace.cpp)
 FetchContent_Declare(json URL https://github.com/nlohmann/json/releases/download/v3.11.2/json.tar.xz)
 FetchContent_MakeAvailable(json)
 include(FetchContent)
