+++ winter/Code/sys/sys.hpp
@@ -1,15 +1,12 @@
 #pragma once
-#include <adastral_defs.h>
+#include <adastral_defs.hpp>
 #include <filesystem>
-#include <fstream>
-#include <iostream>
 #include <string>
 #include <vdf_parser.hpp>
 #include <vector>
 #include "zip/zip.h"
 
 #ifdef _WIN32
-#define WIN32_LEAN_AND_MEAN
 #include <Windows.h>
 #include <Winreg.h>
 #endif
