+++ winter/Code/torrent/torrent.cpp
@@ -1,4 +1,3 @@
-
 #include <torrent.hpp>
 
 namespace
